Moved enums.py inside search.py
* Also added filter_execution param to tests.
